+++ packages/fluentui/docs/src/components/GuidesNavigationFooter.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { Link } from 'react-router-dom';
-import { Button, Flex, Divider } from '@fluentui/react-northstar';
+import { Button, Flex, Divider, Icon } from '@fluentui/react-northstar';
 
 export type PageDescriptor = {
   name: string;
@@ -19,9 +19,25 @@ const GuidesNavigationFooter: React.FC<GuidesNavigationFooterProps> = ({ previou
     <br />
     <Flex gap="gap.small">
       {previous && (
-        <Button as={Link} content={previous.name} icon="arrow left" iconPosition="before" primary to={previous.url} />
+        <Button
+          as={Link}
+          content={previous.name}
+          icon={<Icon name="arrow left" />}
+          iconPosition="before"
+          primary
+          to={previous.url}
+        />
       )}
-      {next && <Button as={Link} content={next.name} icon="arrow right" iconPosition="after" primary to={next.url} />}
+      {next && (
+        <Button
+          as={Link}
+          content={next.name}
+          icon={<Icon name="arrow right" />}
+          iconPosition="after"
+          primary
+          to={next.url}
+        />
+      )}
     </Flex>
   </>
 );