+++ packages/fluentui/docs/src/examples/components/Avatar/Variations/AvatarExampleImageCustomization.shorthand.tsx
@@ -1,16 +1,17 @@
 import * as React from 'react';
 import { Avatar, Icon } from '@fluentui/react-northstar';
+import { AcceptIcon } from '@fluentui/react-icons-northstar';
 
 const AvatarExampleImageCustomizationShorthand = () => (
   <>
     <Avatar
       image={{ src: 'public/images/avatar/small/matt.jpg', alt: 'Profile picture of John Doe' }}
-      status={{ color: 'green', icon: 'icon-checkmark', title: 'Available' }}
+      status={{ color: 'green', icon: <AcceptIcon />, title: 'Available' }}
     />
     &emsp;
     <Avatar
       image="public/images/avatar/large/jerry.png"
-      status={{ color: 'green', icon: 'icon-checkmark', title: 'Available' }}
+      status={{ color: 'green', icon: <AcceptIcon />, title: 'Available' }}
     />
     &emsp;
     <Avatar
@@ -29,7 +30,7 @@ const AvatarExampleImageCustomizationShorthand = () => (
           />
         ),
       }}
-      status={{ color: 'green', icon: 'icon-checkmark', title: 'Available' }}
+      status={{ color: 'green', icon: <AcceptIcon />, title: 'Available' }}
     />
   </>
 );