+++ packages/fluentui/docs/src/examples/components/Popup/Variations/PopupExamplePosition.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import * as _ from 'lodash';
-import { Button, Grid, Popup, Alignment, Position } from '@fluentui/react-northstar';
+import { Button, Grid, Popup, Alignment, Position, Icon } from '@fluentui/react-northstar';
 import { useBooleanKnob, useSelectKnob } from '@fluentui/docs-components';
 
 const PopupExamplePosition = () => {
@@ -32,7 +32,7 @@ const PopupExamplePosition = () => {
         }
       >
         <Button
-          icon={{ name: iconNames[position], circular: true, bordered: true }}
+          icon={<Icon {...{ name: iconNames[position], circular: true, bordered: true }} />}
           styles={buttonStyles}
           title="Show popup"
         />