+++ packages/fluentui/docs/src/prototypes/chatMessages/ChatMessageWithPopover/Popover.tsx
@@ -1,4 +1,5 @@
-import { Accessibility, Menu, menuAsToolbarBehavior } from '@fluentui/react-northstar';
+import { Accessibility, Menu, Icon, menuAsToolbarBehavior } from '@fluentui/react-northstar';
+import { TranslationIcon } from '@fluentui/react-icons-northstar';
 import * as React from 'react';
 import cx from 'classnames';
 
@@ -53,35 +54,35 @@ class Popover extends React.Component<PopoverProps, PopoverState> {
         className={cx(this.props.className, this.state.focused ? 'focused' : '')}
         items={[
           {
+            icon: <Icon name="smile" />,
             key: 'smile',
-            icon: 'smile',
             className: 'smile-emoji',
             'aria-label': 'smile one',
             onClick: this.handleActionableItemClick,
           },
           {
             key: 'smile2',
-            icon: 'smile',
+            icon: <Icon name="smile" />,
             className: 'smile-emoji',
             'aria-label': 'smile two',
             onClick: this.handleActionableItemClick,
           },
           {
             key: 'smile3',
-            icon: 'smile',
+            icon: <Icon name="smile" />,
             className: 'smile-emoji',
             'aria-label': 'smile three',
             onClick: this.handleActionableItemClick,
           },
           {
             key: 'a',
-            icon: 'thumbs up',
+            icon: <Icon name="thumbs up" />,
             'aria-label': 'thumbs up',
             onClick: this.handleActionableItemClick,
           },
           {
             key: 'c',
-            icon: 'ellipsis horizontal',
+            icon: <Icon name="ellipsis horizontal" />,
             onMenuOpenChange: (e, { menuOpen }) => {
               onShowActionMenuChange(true);
               onForceShowActionMenuChange(menuOpen);
@@ -91,9 +92,9 @@ class Popover extends React.Component<PopoverProps, PopoverState> {
             menu: {
               pills: true,
               items: [
-                { key: 'bookmark', icon: 'folder', content: 'Save this message' },
-                { key: 'linkify', icon: 'linkify', content: 'Copy link' },
-                { key: 'translate', icon: 'translate', content: 'Translate' },
+                { key: 'bookmark', icon: <Icon name="folder" />, content: 'Save this message' },
+                { key: 'linkify', icon: <Icon name="linkify" />, content: 'Copy link' },
+                { key: 'translate', icon: <TranslationIcon />, content: 'Translate' },
               ],
             },
           },